project/requirements.txt
asyncpg==0.21.0
black==20.8b1
fastapi==0.61.1
flake8===3.8.4
gunicorn==20.0.4
isort==4.3.21
pytest==6.1.2
pytest-cov==2.10.1
requests==2.24.0
tortoise-orm==0.16.17
uvicorn[standard]==0.12.2